@@ -5,12 +5,13 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:flutter_easyloading/flutter_easyloading.dart';
 import 'package:miko/model/dictionary_model.dart';
-import 'package:miko/page/chat/widget.dart';
 import 'package:miko/page/dictionary/dictionary_view_model.dart';
 import 'package:miko/page/setting/setting_view_model.dart';
 import 'package:miko/utils/app_utils.dart';
 import 'package:miko/utils/routes.dart';
 import 'package:provider/provider.dart';
+
+import '../page/chat/widget.dart';
 
 class Message {
   String name = '';
